refactor: move readStdin to parseString
Issue: BTC-1351
